+++ diary-api/src/main/java/chungha/diary/model/request/DiaryDeleteReq.java
@@ -1,6 +1,6 @@
 package chungha.diary.model.request;
 
 public record DiaryDeleteReq(
-        String diaryId
+	String diaryId
 ) {
 }
