+++ diary-api/src/main/java/chungha/diary/model/request/DiaryUpdateReq.java
@@ -3,9 +3,9 @@ package chungha.diary.model.request;
 import chungha.diarycommon.model.Emotion;
 
 public record DiaryUpdateReq(
-        String diaryId,
-        String title,
-        String content,
-        Emotion emotion
+	String diaryId,
+	String title,
+	String content,
+	Emotion emotion
 ) {
 }
