+++ diary-llm/src/main/java/com/chungha/diaryllm/config/MongoDBConfig.java
@@ -9,16 +9,16 @@ import org.springframework.data.mongodb.core.SimpleMongoClientDatabaseFactory;
 
 @Configuration
 public class MongoDBConfig {
-    @Value("${spring.data.mongodb.uri}")
-    private String connectionString;
+	@Value("${spring.data.mongodb.uri}")
+	private String connectionString;
 
-    @Bean
-    public MongoDatabaseFactory mongoDatabaseFactory() {
-        return new SimpleMongoClientDatabaseFactory(connectionString);
-    }
+	@Bean
+	public MongoDatabaseFactory mongoDatabaseFactory() {
+		return new SimpleMongoClientDatabaseFactory(connectionString);
+	}
 
-    @Bean
-    public MongoTemplate mongoTemplate() {
-        return new MongoTemplate(mongoDatabaseFactory());
-    }
+	@Bean
+	public MongoTemplate mongoTemplate() {
+		return new MongoTemplate(mongoDatabaseFactory());
+	}
 }
