+++ diary-llm/src/main/java/com/chungha/diaryllm/model/request/FeedbackReq.java
@@ -1,6 +1,7 @@
 package com.chungha.diaryllm.model.request;
 
 public record FeedbackReq(
-        String content
+	String diaryId,
+	String content
 ) {
 }
