+++ diary-llm/src/main/java/com/chungha/diaryllm/model/response/FeedbackRes.java
@@ -1,4 +1,5 @@
 package com.chungha.diaryllm.model.response;
 
-public record FeedbackRes(String message) {}
+public record FeedbackRes(String message) {
+}
